minor modification to S.4.4.
Section 4.4 had said "If you are unsure whom to assign as a reviewer, ask the community by posting on the Rubin Community Forum under the Support category." but this is incorrect. Now says "If you are unsure whom to assign as a reviewer, ask the Lead Community Scientist to help identify someone".